kook-924 добавление мер продуктам, закрыть #924

+++ kook/views/product.py
@@ -8,30 +8,37 @@ def index(request):
             'units': Unit.fetch_all()}
 
 def update(request):
-    title = request.matchdict['title']
+    if 'title' in request.matchdict:
+        title = request.matchdict['title']
+        product = Product.fetch(title) or Product(title)
+    else:
+        title = ''
+        product = Product.dummy()
     try:
         update_path = request.current_route_url(title=title)
         next = request.route_path('products')
     except ValueError:
         update_path = '/'
         next = '/'
-    product = Product.fetch(title)
     if request.POST:
         new_title = request.POST.getone('title')
-        existing_product = Product.fetch(new_title)
-        if existing_product:
-            for ingredient in Ingredient.fetch_all(product_title=title):
-                ingredient.product = existing_product
-            product.delete()
-            request.session.flash(u'<div class="alert alert-success">'
-                                  u'Продукт "%s" заменен продуктом "%s"</div>'
-                                  % (title, new_title))
-        else:
-            product.title = new_title
-            request.session.flash(u'<div class="alert alert-success">'
-                                  u'Продукт "%s" обновлен</div>' % product.title)
+        if new_title != product.title:
+            existing_product = Product.fetch(new_title)
+            if existing_product:
+                for ingredient in Ingredient.fetch_all(product_title=title):
+                    ingredient.product = existing_product
+                product.delete()
+                request.session.flash(u'<div class="alert alert-success">'
+                                      u'Продукт "%s" заменен продуктом "%s"</div>'
+                                    % (title, new_title))
+                return HTTPFound(next)
+        product.update_from_multidict(request.POST)
+        product.save()
+        request.session.flash(u'<div class="alert alert-success">'
+                              u'Продукт "%s" обновлен</div>' % product.title)
         return HTTPFound(next)
-    return {'product':product, 'update_path':update_path}
+    return {'product':product, 'update_path':update_path,
+            'units': Unit.fetch_all()}
 
 def delete(request):
     title = request.matchdict['title']
